Moved tests to subproject folders
